+++ bigbang/analysis/listserv.py
@@ -21,7 +21,11 @@ from bs4 import BeautifulSoup
 
 from config.config import CONFIG
 
-from bigbang.io import ListservMessageIO, ListservListIO, ListservArchiveIO
+from bigbang.bigbang_io import (
+    ListservMessageIO,
+    ListservListIO,
+    ListservArchiveIO,
+)
 from bigbang.utils import (
     get_paths_to_files_in_directory,
     get_paths_to_dirs_in_directory,
@@ -265,8 +269,7 @@ class ListservList:
         -------
         `ListservList` object cropped to specification.
         """
-        mlist = self.df.dropna(subset=['from'])
-        generator = ListservList.iterator_name_localpart_domain(mlist["from"].values)
+        mlist = self.df.dropna(subset=["from"])
         if "domain" in list(per_address_field.keys()):
             _addr = pd.Series(
                 [
@@ -278,8 +281,15 @@ class ListservList:
             mlist = mlist.loc[_addr.index]
             mlist = mlist[_addr.isin(per_address_field["domain"])]
         if "localpart" in list(per_address_field.keys()):
-            _addr = pd.Series([localpart for _, localpart, _ in generator])
-            indices = _addr[_addr.isin(per_address_field["localpart"])].index.values
+            _addr = pd.Series(
+                [
+                    ListservList.get_name_localpart_domain(addr)[1]
+                    for addr in mlist["from"].values
+                ],
+                index=mlist.index.values,
+            ).dropna()
+            mlist = mlist.loc[_addr.index]
+            mlist = mlist[_addr.isin(per_address_field["localpart"])]
         return ListservList.from_pandas_dataframe(
             df=mlist,
             name=self.name,
@@ -896,7 +906,7 @@ class ListservList:
         self,
         nw: Optional[dict] = None,
         entity_in_focus: Optional[list] = None,
-        node_attributes: Optional[Dict[str,  list]] = None,
+        node_attributes: Optional[Dict[str, list]] = None,
     ):
         """
         Create directed graph from messaging network created with
@@ -948,10 +958,14 @@ class ListservList:
                     node_attributes["node_name"] == node
                 ]
                 if any(attribute_row.index):
-                    node_attribute_value[node] = attribute_row["node_attribute"].values[0]
+                    node_attribute_value[node] = attribute_row[
+                        "node_attribute"
+                    ].values[0]
                 else:
                     node_attribute_value[node] = "-"
-            nx.set_node_attributes(DG, node_attribute_value, name="node_attribute")
+            nx.set_node_attributes(
+                DG, node_attribute_value, name="node_attribute"
+            )
         # attach directed graph to class
         self.dg = DG
